+++ content/team/brian-fletcher/index.md
@@ -1,0 +1,9 @@
+---
+name: Brian Fletcher
+role: Software Engineer
+headshot: './roadie-hoodie.jpg'
+bio: Brian should write about himself here.
+twitterUrl: https://twitter.com/punkle
+linkedinUrl: https://www.linkedin.com/in/brianjffletcher/
+githubUrl: https://github.com/punkle
+---
