+++ content/team/iain-billett/index.md
@@ -1,0 +1,8 @@
+---
+name: 'Iain Billett'
+role: 'Software Engineer'
+headshot: './gray-top.jpg'
+bio: 'Iain should write about himself here.'
+linkedinUrl: 'https://www.linkedin.com/in/iain-billett-0642b531/'
+githubUrl: 'https://github.com/iain-b'
+---
