+++ content/team/irma-solakovic/index.md
@@ -1,0 +1,8 @@
+---
+name: 'Irma Solakovic'
+role: 'Software Engineer'
+headshot: './light-blue-top.jpeg'
+bio: 'Irma should write about herself here.'
+linkedinUrl: 'https://www.linkedin.com/in/irma-solakovic-70b13b9a/'
+githubUrl: 'https://github.com/Irma12'
+---
