+++ content/team/jussi-hallila/index.md
@@ -1,0 +1,9 @@
+---
+name: Jussi Hallila
+role: Software Engineer
+headshot: './roadie-hoodie.jpg'
+bio: Jussi should write about himself here.
+twitterUrl: https://twitter.com/Xantier
+linkedinUrl: https://www.linkedin.com/in/jussihallila/
+githubUrl: https://github.com/Xantier
+---
