+++ content/team/orla-tuite/index.md
@@ -1,0 +1,9 @@
+---
+name: Orla Tuite
+role: Chief of Staff
+headshot: './smile.jpg'
+bio: Orla should write about herself here.
+twitterUrl: https://twitter.com/OrlaChewit
+linkedinUrl: https://www.linkedin.com/in/orla-t-01a80682/
+githubUrl: https://github.com/oblongcoffee
+---
