+++ content/team/sam-blausten/index.md
@@ -1,0 +1,9 @@
+---
+name: Sam Blausten
+role: Software Engineer
+headshot: './dark-shirt.jpg'
+bio: Sam should write about himself here.
+twitterUrl: 'https://twitter.com/SBlausten'
+linkedinUrl: 'https://www.linkedin.com/in/samblausten/'
+githubUrl: 'https://github.com/sblausten'
+---
