+++ src/components/landing/index.js
@@ -4,5 +4,5 @@ export { default as FooterCTA } from './FooterCTA';
 export { default as CustomerLogoCloud } from './CustomerLogoCloud';
 export { default as VideoHero } from './VideoHero/VideoHero';
 export { default as SimpleCenteredHeading } from './SimpleCenteredHeading';
-export { Team } from './Team';
+export { CircleThreeColTeam } from './Team';
 export { default as ExplainingTheVision } from './ExplainingTheVision';
